+++ Assignment3/src/MapReduce/TagContextWordMapper.java
@@ -2,14 +2,14 @@ package MapReduce;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
 public class TagContextWordMapper extends
-		Mapper<Text, FloatWritable, Text, FloatWritable> {
+		Mapper<Text, DoubleWritable, Text, DoubleWritable> {
 	@Override
-    protected void map(Text key, FloatWritable value, Context context)
+    protected void map(Text key, DoubleWritable value, Context context)
     		throws IOException, InterruptedException
 	{
 		context.write(key, value);
