+++ Assignment3/src/MapReduce/WordInfoReducer.java
@@ -9,12 +9,12 @@ package MapReduce;
 
 
 import java.io.IOException;
-import java.io.PrintStream;
+//import java.io.PrintStream;
 import java.util.ArrayList;
 // import org.apache.commons.logging.Log;
 // import org.apache.commons.logging.LogFactory;
 import org.apache.hadoop.mapreduce.Reducer;
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.LongWritable;
 
@@ -22,7 +22,7 @@ import org.apache.hadoop.io.LongWritable;
  *
  * @author yarneo
  */
-public class WordInfoReducer extends Reducer<Text,Pair,Text,FloatWritable> {
+public class WordInfoReducer extends Reducer<Text,Pair,Text,DoubleWritable> {
     // The Karmasphere Studio Workflow Log displays logging from Apache Commons Logging, for example:
     // private static final Log LOG = LogFactory.getLog("MapReduce.ContextInfoReducer");
 
@@ -38,13 +38,13 @@ public class WordInfoReducer extends Reducer<Text,Pair,Text,FloatWritable> {
         	//Key: context /tab word   Value: count(word in context) / count(word) = p(context|word)
         	//String outStr = val.getText().toString() + "-,-" + key.toString();
         	String outStr =key.toString() + "-,-" + val.getText().toString();
-        	float outFloat = ((float)val.getNum().get() / count_of_word);
+        	double outFloat = ((double)val.getNum().get() / count_of_word);
 //        	 String unicodeMessage =
 //        		 "WORDINFOREDUCER: " + outStr + " " + outFloat;
 //
 //        		    PrintStream out = new PrintStream(System.out, true, "UTF-8");
 //        		    out.println(unicodeMessage);
-        	context.write(new Text(outStr), new FloatWritable(outFloat));
+        	context.write(new Text(outStr), new DoubleWritable(outFloat));
         }
     }
 }
